+++ src/components/button/button.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
-import checkHrefAndOnClick from '../../services/prop-types/check_href_and_onclick';
+import checkHrefAndOnClick from '../../services/prop_types/check_href_and_onclick';
 
 import {
   ICON_TYPES,
@@ -32,15 +32,6 @@ const iconSideToClassNameMap = {
 };
 
 export const ICON_SIDES = Object.keys(iconSideToClassNameMap);
-
-// const checkHrefAndOnClick = (props, propName, componentName) => {
-//   if (props.href && props.onClick) {
-//     throw new Error(
-//       `${componentName} must either specify an href property (if it should be a link) ` +
-//       `or an onClick property (if it should be a button), but not both.`
-//     );
-//   }
-// };
 
 export const EuiButton = ({
   children,
@@ -84,7 +75,6 @@ export const EuiButton = ({
   if (href) {
     return (
       <a
-        disabled={isDisabled}
         className={classes}
         href={href}
         {...rest}